Add IsPivot trait.
